codegen: Rename *_layer_func_table to include "vulkan"
