+++ android/layer/CMakeLists.txt
@@ -17,8 +17,8 @@ target_sources(VkLayer_gfxreconstruct
                    ${GFXRECON_SOURCE_DIR}/layer/dll_main.cpp
                    ${GFXRECON_SOURCE_DIR}/layer/trace_layer.h
                    ${GFXRECON_SOURCE_DIR}/layer/trace_layer.cpp
-                   ${GFXRECON_SOURCE_DIR}/framework/encode/custom_layer_func_table.h
-                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_layer_func_table.h
+                   ${GFXRECON_SOURCE_DIR}/framework/encode/custom_vulkan_layer_func_table.h
+                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_vulkan_layer_func_table.h
               )
 
 target_include_directories(VkLayer_gfxreconstruct
